+++ test/perf.js
@@ -2,7 +2,8 @@
 'use strict'    // eslint-disable-line
 
 var
-  tmpl23 = require('../dist/tmpl').tmpl,
+  version = require('../package.json').version,
+  tmplNew = require('../dist/tmpl').tmpl,
   tmpl22 = require('./v223/tmpl223.js').tmpl
 
 var
@@ -32,7 +33,7 @@ var
     ['{ foo: str + num }', 'foo']
   ],
   ex22a, ex22b, ex22c, mem22, tt22 = [],
-  ex23a, ex23b, ex23c, mem23, tt23 = []
+  exNewa, exNewb, exNewc, memNew, tt23 = []
 
 var LOOP = 50000, TMAX = 12, CPAD = 12, NPAD = 11
 
@@ -48,27 +49,27 @@ ex22b = tt22.reduce(numsum)
 testExpr(tmpl22, data, tt22, tmplList, mem22)
 ex22c = tt22.reduce(numsum)
 
-console.log('tmpl v2.3.0 ...')
-mem23 = [0, 0, 0]
-testExpr(tmpl23, data, tt23, exprList, mem23, 1)
-ex23a = tt23.reduce(numsum)
-testExpr(tmpl23, data, tt23, csList, mem23, 1)
-ex23b = tt23.reduce(numsum)
-testExpr(tmpl23, data, tt23, tmplList, mem23, 1)
-ex23c = tt23.reduce(numsum)
+console.log('tmpl v' + version + ' ...')
+memNew = [0, 0, 0]
+testExpr(tmplNew, data, tt23, exprList, memNew, 1)
+exNewa = tt23.reduce(numsum)
+testExpr(tmplNew, data, tt23, csList, memNew, 1)
+exNewb = tt23.reduce(numsum)
+testExpr(tmplNew, data, tt23, tmplList, memNew, 1)
+exNewc = tt23.reduce(numsum)
 
 console.log()
-console.log('%s    tmpl 2.2.4   new v2.3.0', padr('Results', CPAD))
+console.log('%s    tmpl 2.2.4   new v' + version, padr('Results', CPAD))
 console.log('%s    ----------   ----------', replicate('-', CPAD))
-console.log('%s:  %s %s', padr('Expressions', CPAD), padl(ex22a, NPAD), padl(ex23a, NPAD))
-console.log('%s:  %s %s', padr('Shorthands',  CPAD), padl(ex22b, NPAD), padl(ex23b, NPAD))
-console.log('%s:  %s %s', padr('Templates',   CPAD), padl(ex22c, NPAD), padl(ex23c, NPAD))
-console.log('%s:  %s %s', padr('TOTAL',       CPAD), padl(ex22a + ex22b + ex22c, NPAD), padl(ex23a + ex23b + ex23c, NPAD))
+console.log('%s:  %s %s', padr('Expressions', CPAD), padl(ex22a, NPAD), padl(exNewa, NPAD))
+console.log('%s:  %s %s', padr('Shorthands',  CPAD), padl(ex22b, NPAD), padl(exNewb, NPAD))
+console.log('%s:  %s %s', padr('Templates',   CPAD), padl(ex22c, NPAD), padl(exNewc, NPAD))
+console.log('%s:  %s %s', padr('TOTAL',       CPAD), padl(ex22a + ex22b + ex22c, NPAD), padl(exNewa + exNewb + exNewc, NPAD))
 console.log()
 console.log('Memory')
-//console.log('%s:  %s %s', padr('Res set size', CPAD), padl(mem22[0], NPAD), padl(mem23[0], NPAD))
-console.log('%s:  %s %s', padr('Heap total',   CPAD), padl(mem22[1], NPAD), padl(mem23[1], NPAD))
-console.log('%s:  %s %s', padr('Heap used',    CPAD), padl(mem22[2], NPAD), padl(mem23[2], NPAD))
+//console.log('%s:  %s %s', padr('Res set size', CPAD), padl(mem22[0], NPAD), padl(memNew[0], NPAD))
+console.log('%s:  %s %s', padr('Heap total',   CPAD), padl(mem22[1], NPAD), padl(memNew[1], NPAD))
+console.log('%s:  %s %s', padr('Heap used',    CPAD), padl(mem22[2], NPAD), padl(memNew[2], NPAD))
 console.log()
 
 console.log('NOTES:')
@@ -95,8 +96,9 @@ function testExpr (tmpl, data, times, list, agc) {
 
     s = tmpl(expr, data)
     //assert(s === pair[1])
-    if (s !== pair[1])
+    if (s !== pair[1]) {
       throw new Error('`' + s + '` in #' + idx  + ' is not `' + pair[1] + '`')
+    }
 
     for (i = 0; i < tt.length; ++i) {
       tt[i] = Date.now()